UT: Move event_queue_subscriber fixture back to test_pypubsub_event_queue.py
